embed web resources
